refactor(core): extract region validation + resolution into maestro-regions.js
Move validateRegionInputs (shape checks) and resolveRegions (element/coordinate
bbox resolution → comparison-payload fragments) out of handleMaestroScreenshot
into a dedicated module. The hierarchy dump and warn-once flag stay
request-scoped (call-local inside resolveRegions), preserving the per-request
memoization invariant. Three independent output fields, algorithm pass-through,
and percy.grpcClientCache threading are unchanged. istanbul-ignore annotations
travel verbatim. No behavior change.

maestro-screenshot.js is now a 182-line orchestrator.
