+++ packages/core/src/maestro-screenshot.js
@@ -4,8 +4,8 @@ import { normalize } from '@percy/config/utils';
 import { ServerError } from './server.js';
 import { encodeURLSearchParams } from './utils.js';
 import { handleSyncJob } from './snapshot.js';
-import { dump as maestroDump, firstMatch as maestroFirstMatch, SELECTOR_KEYS_WHITELIST } from './maestro-hierarchy.js';
 import { locateScreenshot } from './maestro-screenshot-file.js';
+import { validateRegionInputs, resolveRegions } from './maestro-regions.js';
 
 // Parse PNG IHDR chunk for the screenshot's actual rendered dimensions.
 // Returns { width, height } when the buffer is a valid PNG with non-zero
@@ -85,42 +85,8 @@ export async function handleMaestroScreenshot(req, res, percy) {
   }
 
   // Validate regions input shape early (before file I/O and ADB work) so
-  // malformed requests don't consume resolver/relay work. Three parallel
-  // input arrays share the same per-item shape; algorithm semantics differ
-  // per array (regions only — ignoreRegions/considerRegions are implicit).
-  const REGION_INPUT_FIELDS = ['regions', 'ignoreRegions', 'considerRegions'];
-  for (let fieldName of REGION_INPUT_FIELDS) {
-    let input = req.body[fieldName];
-    if (input === undefined) continue;
-    if (!Array.isArray(input)) {
-      throw new ServerError(400, `${fieldName} must be an array`);
-    }
-    if (input.length > 50) {
-      throw new ServerError(400, `${fieldName} exceeds maximum of 50`);
-    }
-    for (let [idx, region] of input.entries()) {
-      if (region && region.element !== undefined) {
-        if (typeof region.element !== 'object' || region.element === null || Array.isArray(region.element)) {
-          throw new ServerError(400, `${fieldName}[${idx}].element must be an object`);
-        }
-        let keys = Object.keys(region.element);
-        if (keys.length !== 1) {
-          throw new ServerError(400, `${fieldName}[${idx}].element must have exactly one selector key`);
-        }
-        let [key] = keys;
-        if (!SELECTOR_KEYS_WHITELIST.includes(key)) {
-          throw new ServerError(400, `${fieldName}[${idx}].element: unsupported selector key "${key}" (allowed: ${SELECTOR_KEYS_WHITELIST.join(', ')})`);
-        }
-        let value = region.element[key];
-        if (typeof value !== 'string' || value.length === 0) {
-          throw new ServerError(400, `${fieldName}[${idx}].element.${key} must be a non-empty string`);
-        }
-        if (value.length > 512) {
-          throw new ServerError(400, `${fieldName}[${idx}].element.${key} exceeds maximum length of 512`);
-        }
-      }
-    }
-  }
+  // malformed requests don't consume resolver/relay work.
+  validateRegionInputs(req.body);
 
   // Locate the screenshot on disk (supplied filePath or legacy glob) and
   // confirm it resolves under the sessionId-owned dir. Throws ServerError(404)
@@ -172,196 +138,11 @@ export async function handleMaestroScreenshot(req, res, percy) {
   if (req.body.labels) payload.labels = req.body.labels;
   if (req.body.thTestCaseExecutionId) payload.thTestCaseExecutionId = req.body.thTestCaseExecutionId;
 
-  // ───────────────────────────────────────────────────────────────────
-  // REGIONS — end-to-end architecture
-  // ───────────────────────────────────────────────────────────────────
-  //
-  // Regions tell Percy's diff engine which parts of a mobile screenshot
-  // to ignore / consider / layout-compare. Two ways to specify one:
-  //
-  //   1. Coordinate region — caller already knows the pixel rectangle.
-  //      Shape: { top, left, right, bottom }. Forwarded as-is after
-  //      transform to `{x, y, width, height}` boundingBox.
-  //
-  //   2. Element region — caller knows a selector (`resource-id`, `text`,
-  //      `content-desc`, `class`, `id`) but not the on-screen bounds.
-  //      Resolved at relay-time against the live device's view hierarchy.
-  //
-  // ── Data flow (element region case) ────────────────────────────────
-  //
-  //   SDK (percy-screenshot.js)
-  //     │  POST /percy/maestro-screenshot
-  //     │   { name, sessionId, platform, regions:[{element:{...}}], ... }
-  //     ▼
-  //   Relay (this handler)
-  //     │  validate selector shape (SELECTOR_KEYS_WHITELIST)
-  //     │  maestroDump({ platform, sessionId, grpcClientCache })  ← lazy + memoized per request
-  //     │      │
-  //     │      ├─ Android cascade (maestro-hierarchy.js)
-  //     │      │    gRPC primary → maestro-CLI → adb uiautomator
-  //     │      │    Three-class taxonomy: schema-class (drift bit, no
-  //     │      │    fallback) / channel-broken (evict cache, fall back) /
-  //     │      │    contention-class (keep cache, skip CLI → adb).
-  //     │      │
-  //     │      └─ iOS cascade
-  //     │           HTTP primary (Maestro XCTestRunner /viewHierarchy)
-  //     │           → maestro-CLI shell-out. AUT-root detection skips
-  //     │           SpringBoard frames.
-  //     │
-  //     │  firstMatch(nodes, selector) → bbox or null (warn-skip).
-  //     │  payload.regions[i].elementSelector.boundingBox = bbox
-  //     ▼
-  //   Percy backend — compares masked regions across builds.
-  //
-  // ── Observability ──────────────────────────────────────────────────
-  //
-  // /percy/healthcheck exposes maestroHierarchyDrift per platform:
-  //   { lastFailureClass, fallbackCount, succeededVia, code?, reason?, firstSeenAt? }
-  // Every primary→fallback transition also emits one info-level line:
-  //   [percy] hierarchy: <primary> failed (<class>: <reason>) → falling back to <next>
-  //
-  // ── Failure shape ──────────────────────────────────────────────────
-  //
-  // Element regions degrade gracefully: resolver failure → warn-skip
-  // those regions only; the snapshot itself still uploads. Coordinate
-  // regions don't depend on the resolver and always pass through.
-  //
-  // ───────────────────────────────────────────────────────────────────
-  // Shared resolver state across regions/ignoreRegions/considerRegions —
-  // one hierarchy dump per request, one warn-once skip notice.
-  let cachedDump = null;
-  let elementSkipWarned = false;
-  const totalElementRegionCount = REGION_INPUT_FIELDS.reduce((sum, f) => {
-    let arr = req.body[f];
-    return sum + (Array.isArray(arr) ? arr.filter(r => r && r.element).length : 0);
-  }, 0);
-
-  // Resolve one region input to {x, y, width, height}, or null when the
-  // region is invalid or the resolver couldn't match it. Mutates the
-  // shared cachedDump / warn-flag state above.
-  async function resolveBbox(region) {
-    if (region.top != null && region.bottom != null && region.left != null && region.right != null) {
-      return {
-        x: region.left,
-        y: region.top,
-        width: region.right - region.left,
-        height: region.bottom - region.top
-      };
-    }
-    /* istanbul ignore else — region.element false branch falls through
-       to the istanbul-ignored "Invalid region format" warn below. */
-    if (region.element) {
-      /* istanbul ignore else — cachedDump === null only on first
-         element-region per request; subsequent regions hit the cache. */
-      if (cachedDump === null) {
-        // Thread the per-Percy gRPC client cache so the Android gRPC
-        // primary path can reuse channels across snapshots in the same
-        // session (D9 of 2026-05-07-002 plan). iOS path ignores it.
-        cachedDump = await maestroDump({
-          platform,
-          sessionId,
-          grpcClientCache: percy.grpcClientCache
-        });
-      }
-      /* istanbul ignore else — branch where dump resolves to hierarchy is
-         happy-path element-region territory, integration-tested only. */
-      if (cachedDump.kind !== 'hierarchy') {
-        /* istanbul ignore else — elementSkipWarned latches after first
-           warn; second+ iterations take the no-op branch. */
-        if (!elementSkipWarned) {
-          percy.log.warn(
-            `Element-region resolver ${cachedDump.kind} (${cachedDump.reason}) — skipping ${totalElementRegionCount} element regions`
-          );
-          elementSkipWarned = true;
-        }
-        return null;
-      }
-      /* istanbul ignore next */
-      let bbox = maestroFirstMatch(cachedDump.nodes, region.element);
-      /* istanbul ignore next */
-      if (!bbox) {
-        percy.log.warn(`Element region not found: ${JSON.stringify(region.element)} — skipping`);
-        return null;
-      }
-      /* istanbul ignore next — element-region happy path requires a
-         non-stub maestroDump returning hierarchy nodes; unit tests run
-         with stubbed resolver (env-missing), happy path covered by the
-         cross-platform-parity integration harness against fixture data. */
-      return bbox;
-    }
-    /* istanbul ignore next */
-    percy.log.warn('Invalid region format, skipping');
-    /* istanbul ignore next — region shape is validated upstream by the
-       SDK before posting; this is a defensive catch-all for regions that
-       lack both coordinate fields AND an element selector. */
-    return null;
-  }
-
-  // regions[]: comparison-shape items with algorithm. Default algorithm is
-  // 'ignore' (back-compat with SDK ≤ 0.3).
-  if (Array.isArray(req.body.regions)) {
-    let resolvedRegions = [];
-    for (let region of req.body.regions) {
-      let bbox = await resolveBbox(region);
-      if (!bbox) continue;
-      let resolved = {
-        elementSelector: { boundingBox: bbox },
-        algorithm: region.algorithm || 'ignore'
-      };
-      /* istanbul ignore if — region.configuration optional field; only
-         passed when SDK opts in to per-region config overrides. */
-      if (region.configuration) resolved.configuration = region.configuration;
-      /* istanbul ignore if — region.padding optional field. */
-      if (region.padding) resolved.padding = region.padding;
-      /* istanbul ignore if — region.assertion optional field. */
-      if (region.assertion) resolved.assertion = region.assertion;
-      resolvedRegions.push(resolved);
-    }
-    /* istanbul ignore else — empty resolvedRegions branch only fires when
-       ALL regions failed to resolve; happy path resolves at least one. */
-    if (resolvedRegions.length > 0) payload.regions = resolvedRegions;
-  }
-
-  // ignoreRegions[] and considerRegions[]: parallel top-level payload
-  // fields. Each item is shaped per regionsSchema (config.js:792) —
-  // { coOrdinates: {top, left, bottom, right} } with an optional selector
-  // hint preserved when the caller supplied an element selector.
-  const REGION_OUTPUT_MAP = {
-    ignoreRegions: { payloadKey: 'ignoredElementsData', innerKey: 'ignoreElementsData' },
-    considerRegions: { payloadKey: 'consideredElementsData', innerKey: 'considerElementsData' }
-  };
-  for (let [inputField, { payloadKey, innerKey }] of Object.entries(REGION_OUTPUT_MAP)) {
-    let input = req.body[inputField];
-    if (!Array.isArray(input)) continue;
-    let resolved = [];
-    for (let region of input) {
-      let bbox = await resolveBbox(region);
-      /* istanbul ignore if — null bbox skip in ignoreRegions/considerRegions
-         loop; tests cover the happy path where every region resolves. */
-      if (!bbox) continue;
-      let item = {
-        coOrdinates: {
-          top: bbox.y,
-          left: bbox.x,
-          bottom: bbox.y + bbox.height,
-          right: bbox.x + bbox.width
-        }
-      };
-      /* istanbul ignore if — element selector echo on resolved region;
-         only fires when resolveBbox returned a bbox for an element region,
-         which itself is integration-test territory (see resolveBbox
-         above for the resolver-mock rationale). */
-      if (region.element) {
-        let [key] = Object.keys(region.element);
-        item.selector = `${key}=${region.element[key]}`;
-      }
-      resolved.push(item);
-    }
-    /* istanbul ignore else — empty resolved branch only fires when ALL
-       regions in this category failed to resolve; happy path resolves
-       at least one. */
-    if (resolved.length > 0) payload[payloadKey] = { [innerKey]: resolved };
-  }
+  // Resolve element/coordinate regions to comparison-payload fragments and
+  // merge them into the payload. Element regions degrade gracefully — a
+  // resolver failure warn-skips those regions only; the snapshot still
+  // uploads. The hierarchy dump is memoized per request inside resolveRegions.
+  Object.assign(payload, await resolveRegions({ body: req.body, platform, sessionId, percy }));
 
   // Upload via percy — sync or fire-and-forget
   if (req.body.sync === true) payload.sync = true;
